requirements.txt
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
Flask-Limiter==3.3.0
limits==3.5.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.7.1
PyJWT==2.10.1
python-dotenv==1.0.0
WeasyPrint==59.0
email-validator==2.0.0.post2
mysqlclient==2.2.4
Jinja2==3.1.2
itsdangerous==2.1.2
Werkzeug==2.3.7
gunicorn==20.1.0
pydyf==0.8.0
pdfkit==1.0.0
python-magic==0.4.27
argon2-cffi==23.1.0
pyclamd==0.4.0
pymysql>=1.0.2
